fix: Check hook status after an upgrade/reinstall/uninstall/install (#3195)
Problem:
@mraszyk pointed out
[here](https://github.com/dfinity/portal/pull/3761#issuecomment-2478612486):

> The hook should not run after an upgrade/reinstall/uninstall/install
if the condition is not satisfied after the upgrade (although the
condition was satisfied before the upgrade and the hook did not execute
yet).

Solution:
This PR should enforce that the status of the hook condition is updated
each time we `upgrade/reinstall/uninstall/install` code.
