+++ rs/execution_environment/src/execution_environment/tests/canister_task.rs
@@ -2,6 +2,7 @@ use assert_matches::assert_matches;
 use ic_config::{execution_environment::Config as HypervisorConfig, subnet_config::SubnetConfig};
 use ic_error_types::RejectCode;
 use ic_management_canister_types::{CanisterSettingsArgsBuilder, CanisterStatusType};
+use ic_management_canister_types::{CanisterUpgradeOptions, WasmMemoryPersistence};
 use ic_registry_subnet_type::SubnetType;
 use ic_replicated_state::canister_state::NextExecution;
 use ic_replicated_state::canister_state::WASM_PAGE_SIZE_IN_BYTES;
@@ -823,6 +824,7 @@ fn global_timer_produces_transient_error_on_out_of_cycles() {
 fn get_wat_with_update_and_hook_mem_grow(
     update_grow_mem_size: i32,
     hook_grow_mem_size: i32,
+    with_enchanced_ortogonal_persistence: bool,
 ) -> String {
     let mut wat = r#"
     (module
@@ -843,7 +845,17 @@ fn get_wat_with_update_and_hook_mem_grow(
     wat.push_str(
         r#")))
     )
-    (memory 1 20)
+    (memory 1 20)"#,
+    );
+    if with_enchanced_ortogonal_persistence {
+        wat.push_str(
+            r#"
+            (@custom "icp:private enhanced-orthogonal-persistence" "")
+            "#,
+        );
+    }
+    wat.push_str(
+        r#"
     )"#,
     );
     wat
@@ -856,7 +868,8 @@ fn on_low_wasm_memory_is_executed() {
     let update_grow_mem_size = 7;
     let hook_grow_mem_size = 5;
 
-    let wat = get_wat_with_update_and_hook_mem_grow(update_grow_mem_size, hook_grow_mem_size);
+    let wat =
+        get_wat_with_update_and_hook_mem_grow(update_grow_mem_size, hook_grow_mem_size, false);
 
     let canister_id = test.canister_from_wat(wat.as_str()).unwrap();
 
@@ -905,7 +918,8 @@ fn on_low_wasm_memory_is_executed_before_message() {
     let update_grow_mem_size = 7;
     let hook_grow_mem_size = 5;
 
-    let wat = get_wat_with_update_and_hook_mem_grow(update_grow_mem_size, hook_grow_mem_size);
+    let wat =
+        get_wat_with_update_and_hook_mem_grow(update_grow_mem_size, hook_grow_mem_size, false);
 
     let canister_id = test.canister_from_wat(wat.as_str()).unwrap();
 
@@ -956,13 +970,177 @@ fn on_low_wasm_memory_is_executed_before_message() {
 }
 
 #[test]
+fn on_low_wasm_memory_is_executed_after_upgrade_if_condition_holds() {
+    let mut test = ExecutionTestBuilder::new().with_manual_execution().build();
+
+    let update_grow_mem_size = 7;
+    let hook_grow_mem_size = 5;
+
+    let wat: String =
+        get_wat_with_update_and_hook_mem_grow(update_grow_mem_size, hook_grow_mem_size, true);
+
+    let canister_id = test.canister_from_wat(wat.as_str()).unwrap();
+
+    test.canister_update_wasm_memory_limit_and_wasm_memory_threshold(
+        canister_id,
+        (20 * WASM_PAGE_SIZE_IN_BYTES as u64).into(),
+        (15 * WASM_PAGE_SIZE_IN_BYTES as u64).into(),
+    )
+    .unwrap();
+
+    // Here we have:
+    // wasm_capacity = wasm_memory_limit = 20 Wasm Pages
+    // wasm_memory_threshold = 15 Wasm Pages
+
+    // Initially wasm_memory.size = 1
+    assert_eq!(
+        test.execution_state(canister_id).wasm_memory.size,
+        NumWasmPages::new(1)
+    );
+
+    test.ingress_raw(canister_id, "grow_mem", vec![]);
+    test.ingress_raw(canister_id, "grow_mem", vec![]);
+
+    // First ingress messages gets executed.
+    // wasm_memory.size = 1 + 7 = 8
+    // wasm_capacity - used_wasm_memory < self.wasm_memory_threshold
+    // Hook condition is triggered.
+    test.execute_slice(canister_id);
+    assert_eq!(
+        test.execution_state(canister_id).wasm_memory.size,
+        NumWasmPages::new(8)
+    );
+
+    let result = test.upgrade_canister_v2(
+        canister_id,
+        wat::parse_str(wat).unwrap(),
+        CanisterUpgradeOptions {
+            skip_pre_upgrade: None,
+            wasm_memory_persistence: Some(WasmMemoryPersistence::Keep),
+        },
+    );
+    assert_eq!(result, Ok(()));
+
+    // Upgrade is executed, and the wasm_memory size is unchanged.
+    // Hook condition is triggered.
+    assert_eq!(
+        test.execution_state(canister_id).wasm_memory.size,
+        NumWasmPages::new(8)
+    );
+
+    // Though we have the second ingress message awaiting to be processed,
+    // hook will be executed first.
+    test.execute_slice(canister_id);
+    assert_eq!(
+        test.execution_state(canister_id).wasm_memory.size,
+        NumWasmPages::new(13)
+    );
+
+    // The second ingress message is executed after the hook.
+    test.execute_slice(canister_id);
+    assert_eq!(
+        test.execution_state(canister_id).wasm_memory.size,
+        NumWasmPages::new(20)
+    );
+}
+
+#[test]
+fn on_low_wasm_memory_is_not_executed_after_upgrade_if_condition_becomes_unsatisfied() {
+    let mut test = ExecutionTestBuilder::new().with_manual_execution().build();
+
+    let update_grow_mem_size = 3;
+    let hook_grow_mem_size = 5;
+
+    let wat: String =
+        get_wat_with_update_and_hook_mem_grow(update_grow_mem_size, hook_grow_mem_size, false);
+
+    let canister_id = test.canister_from_wat(wat.as_str()).unwrap();
+
+    test.canister_update_wasm_memory_limit_and_wasm_memory_threshold(
+        canister_id,
+        (20 * WASM_PAGE_SIZE_IN_BYTES as u64).into(),
+        (15 * WASM_PAGE_SIZE_IN_BYTES as u64).into(),
+    )
+    .unwrap();
+
+    // Here we have:
+    // wasm_capacity = wasm_memory_limit = 20 Wasm Pages
+    // wasm_memory_threshold = 15 Wasm Pages
+
+    // Initially wasm_memory.size = 1
+    assert_eq!(
+        test.execution_state(canister_id).wasm_memory.size,
+        NumWasmPages::new(1)
+    );
+
+    test.ingress_raw(canister_id, "grow_mem", vec![]);
+    test.ingress_raw(canister_id, "grow_mem", vec![]);
+    test.ingress_raw(canister_id, "grow_mem", vec![]);
+
+    // First ingress messages gets executed.
+    // wasm_memory.size = 1 + 3 = 4
+    // wasm_capacity - used_wasm_memory > self.wasm_memory_threshold
+    // Hook condition is not triggered.
+    test.execute_slice(canister_id);
+    assert_eq!(
+        test.execution_state(canister_id).wasm_memory.size,
+        NumWasmPages::new(4)
+    );
+
+    // Second ingress messages gets executed.
+    // wasm_memory.size = 4 + 3 = 7
+    // wasm_capacity - used_wasm_memory < self.wasm_memory_threshold
+    // Hook condition is triggered.
+    test.execute_slice(canister_id);
+    assert_eq!(
+        test.execution_state(canister_id).wasm_memory.size,
+        NumWasmPages::new(7)
+    );
+    println!("canister upgrade");
+
+    let result = test.upgrade_canister_v2(
+        canister_id,
+        wat::parse_str(wat).unwrap(),
+        CanisterUpgradeOptions {
+            skip_pre_upgrade: None,
+            wasm_memory_persistence: None,
+        },
+    );
+    assert_eq!(result, Ok(()));
+    println!("canister upgrade");
+
+    // Upgrade is executed, and the wasm_memory size reset to 1.
+    // Hook condition is not triggered.
+    assert_eq!(
+        test.execution_state(canister_id).wasm_memory.size,
+        NumWasmPages::new(1)
+    );
+
+    // Though the hook was initially scheduled, it is now removed
+    // from queue, and the third ingress message will be executed.
+    test.execute_slice(canister_id);
+    assert_eq!(
+        test.execution_state(canister_id).wasm_memory.size,
+        NumWasmPages::new(4)
+    );
+
+    // There are no messages left to be executed.
+    test.execute_slice(canister_id);
+    assert_eq!(
+        test.execution_state(canister_id).wasm_memory.size,
+        NumWasmPages::new(4)
+    );
+}
+
+#[test]
 fn on_low_wasm_memory_is_executed_once() {
     let mut test = ExecutionTestBuilder::new().build();
 
     let update_grow_mem_size = 7;
     let hook_grow_mem_size = 2;
 
-    let wat = get_wat_with_update_and_hook_mem_grow(update_grow_mem_size, hook_grow_mem_size);
+    let wat =
+        get_wat_with_update_and_hook_mem_grow(update_grow_mem_size, hook_grow_mem_size, false);
 
     let canister_id = test.canister_from_wat(wat.as_str()).unwrap();
