+++ rs/system_api/src/lib.rs
@@ -873,7 +873,7 @@ impl MemoryUsage {
 
                         self.add_execution_memory(execution_bytes, execution_memory_type)?;
 
-                        sandbox_safe_system_state.check_on_low_wasm_memory_hook_condition(
+                        sandbox_safe_system_state.update_status_of_low_wasm_memory_hook_condition(
                             None,
                             self.wasm_memory_limit,
                             self.current_usage,
@@ -900,7 +900,7 @@ impl MemoryUsage {
                 self.current_usage = NumBytes::from(new_usage);
                 self.add_execution_memory(execution_bytes, execution_memory_type)?;
 
-                sandbox_safe_system_state.check_on_low_wasm_memory_hook_condition(
+                sandbox_safe_system_state.update_status_of_low_wasm_memory_hook_condition(
                     Some(reserved_bytes),
                     self.wasm_memory_limit,
                     self.current_usage,
